App/requirements.txt
flask-cors
numpy
newspaper3k
sklearn
pandas
flask
nltk
gunicorn
tensorflow
langdetect
